Add files via upload
A mini analysis on Candy Power Rankings.

Data
The data set is located (incl. a short description) here:
https://github.com/fivethirtyeight/data/tree/master/candy-power-ranking
The data set is provided by FiveThirtyEight under the Creative Commons Attribution 4.0 International license (https://creativecommons.org/licenses/by/4.0/)

Scenario:
The purchasing group of a large supermarket chain wants to expand candy offering. These are store brand candies that we sell along the brand offerings. The idea is to create a brand new product. The team is discussing various options at the moment.
Some prefer cookie-based sweets while others think that it should be gummies. The Divisional Director responsible for purchasing has decided to use a more data-driven approach. He contracted with a market research group to collect data on products in the market and their characteristics and customer sentiment.

The market research data is now available and it is your job to find out which product characteristics drive customer sentiment and subsequently make a recommendation on a new product.